+++ robotbase/src/main/java/com/team2357/frc2023/commands/scoring/AutoScoreLowCommandGroup.java
@@ -1,10 +1,10 @@
 package com.team2357.frc2023.commands.scoring;
 
 import com.team2357.frc2023.Constants;
+import com.team2357.frc2023.commands.claw.ClawOpenCommand;
+import com.team2357.frc2023.commands.claw.CloseClawCommand;
 import com.team2357.frc2023.commands.intake.IntakeArmDeployCommand;
 import com.team2357.frc2023.commands.intake.IntakeRollerReverseCommand;
-import com.team2357.frc2023.commands.scoring.util.CloseClawCommand;
-import com.team2357.frc2023.commands.scoring.util.ClawOpenCommand;
 
 import edu.wpi.first.wpilibj2.command.ParallelRaceGroup;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
