+++ robotbase/src/main/java/com/team2357/frc2023/commands/wrist/WristExtendCommand.java
@@ -1,23 +1,23 @@
-package com.team2357.frc2023.commands.scoring.util;
+package com.team2357.frc2023.commands.wrist;
 
 import com.team2357.frc2023.subsystems.WristSubsystem;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 
-public class WristRetractCommand extends CommandBase {
+public class WristExtendCommand extends CommandBase {
     
-    public WristRetractCommand() {
+    public WristExtendCommand() {
         addRequirements(WristSubsystem.getInstance());
     }
 
     @Override
     public void initialize() {
-        WristSubsystem.getInstance().retract();
+        WristSubsystem.getInstance().extend();
     }
 
     @Override
     public boolean isFinished() {
-        return WristSubsystem.getInstance().isRetracted();
+        return WristSubsystem.getInstance().isExtended();
     }
 
 }
