+++ robotbase/src/main/java/com/team2357/frc2023/commands/wrist/WristRetractCommand.java
@@ -1,23 +1,23 @@
-package com.team2357.frc2023.commands.scoring.util;
+package com.team2357.frc2023.commands.wrist;
 
 import com.team2357.frc2023.subsystems.WristSubsystem;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 
-public class WristExtendCommand extends CommandBase {
+public class WristRetractCommand extends CommandBase {
     
-    public WristExtendCommand() {
+    public WristRetractCommand() {
         addRequirements(WristSubsystem.getInstance());
     }
 
     @Override
     public void initialize() {
-        WristSubsystem.getInstance().extend();
+        WristSubsystem.getInstance().retract();
     }
 
     @Override
     public boolean isFinished() {
-        return WristSubsystem.getInstance().isExtended();
+        return WristSubsystem.getInstance().isRetracted();
     }
 
 }
